refactor: jsc setup
