+++ src/external-rpc/index.ts
@@ -1,3 +1,3 @@
 export * from './types.ts'
-export * from './setup.ts'
+export * from './handle-incoming-msg.ts'
 export * from './utils.ts'
