+++ src/scripts/build-browser.ts
@@ -1,7 +1,7 @@
 import * as esbuild from 'esbuild'
 
 const rslt = await esbuild.build({
-	entryPoints: ['src/external-rpc/index.ts'],
+	entryPoints: ['src/external-rpc/setup-browser.ts'],
 	bundle: true,
 	minify: true,
 	outfile: 'browser/resources/attestor-browser.min.mjs',
